Initial stage of conversion away from WebSQL.
Necessary now that WebSQL is completely gone from Chrome/Chromium,
including via legacy flags.

I don't really want a replacement for Fauxbar which works differently,
like OmniHistoryPlus.  I just want Fauxbar, without WebSQL, regardless
of how bad the WebSQL replacement might be.  As explained on #49, the
replacement could be _very_ naive and still be totally fine.

This first stage is the minimum necessary for me to actually use Fauxbar
again.  It uses a static JSON export of my final WebSQL state in a pure
read-only mode.  There is a shim WebSQL API that just logs attempte SQL
statements, and provides an easy API for hand-converting the most
important SELECT statements.

The next steps are something like this:

1. Use extension messaging so that all access to the data is only done
   by the background page, and never by actual pages.  This eliminates
   all problems concurrent access that would otherwise come up later.

2. Make the data be stored in chrome.storage.local instead of just in
   memory.

3. Add an API for doing writes (insert/update/delete), and hand-convert
   the most important ones.

At this point things should be more or less as they were before WebSQL
went away.  Other things that would be good to do at some point include:

4. Convert the rest of the SQL statements, not just the most important
   ones.

4a. Some like distinct and count will need extra functionality in the
   shim API, but that shouldn't be too hard.

4b. Fix the tags joins.  This is likely harder but still should be
   possible and not too bad.

5. Get rid of `openDb()` and the WebSQL shim layer; eg. its concept of
   "transactions" will, at this point, just be useless boilerplate.

6. Move the data access out of the background page, and into a service
   worker.  This should pave the way for getting rid of the background
   page completely, which in turn paves the way for MV3.

7. Clean things up in general, especially around error handling.  (There
   was a lot of error handling specific to WebSQL errors, and the
   `errors` table, which (IMHO) should all just be tossed in favour of
   logging to the console like normal.)

----

To actually use this current code, it needs to be installed as an
unpacked extension, and the static data put in place.  These are the
steps I used to get the static data:

1. Find the sqlite db file.  It should still be there, even though WebSQL
   support is gone.
    - Under your chrome profile dir (eg.  `~/.config/google-chrome/Profile 1/` on linux).
    - In `databases/chrome-extension_XXXX_0/`, where `XXXX` is the extension id (eg. `hibkhcnpkakjniplpfblaoikiggkopka` if you installed from the chrome app store).
    - My file was just called "`21`".

2. Copy it somewhere safe.
    ```
    mkdir ~/fauxbar-data
    cd ~/fauxbar-data
    cp -ai ~/.config/google-chrome/Profile\ 1/databases/chrome-extension_hibkhcnpkakjniplpfblaoikiggkopka_0/21 fauxbar.sqlite
    ```

3. Extract the data to json with these bash commands.
    ```
    sqlite3 ./fauxbar.sqlite .tables > tables
    sqlite3 ./fauxbar.sqlite .schema > schema
    for i in $(<tables); do sqlite3 ./fauxbar.sqlite '.mode json' 'select * from '"$i"';' > "$i".json; done
    ```

4. Install fauxbar to chrome as an unpacked extension (enable developer mode).

5. Put the data json files into a `data` subdir of the unpacked extension.